+++ anyblok/test_bloks/test_blok14/test.py
@@ -1,0 +1,20 @@
+# This file is a part of the AnyBlok project
+#
+#    Copyright (C) 2018 Jean-Sebastien SUZANNE <jssuzanne@anybox.fr>
+#
+# This Source Code Form is subject to the terms of the Mozilla Public License,
+# v. 2.0. If a copy of the MPL was not distributed with this file,You can
+# obtain one at http://mozilla.org/MPL/2.0/.
+from anyblok.declarations import Declarations
+from anyblok.column import Integer, String
+
+
+register = Declarations.register
+Model = Declarations.Model
+
+
+@register(Model)
+class Test:
+
+    id = Integer(primary_key=True)
+    mode = String()